Refactor to reduce type duplication
This is a quick followup to https://github.com/softprops/recap/pull/29 to reduce the duplication of the tuple type `(&'a str, &'a str, Option<&'a FieldOptions>)` by just using `Val<'a>`, which is a struct with the same contents anyway.

Verified by checking that the test still pass.

This doesn't change the public API so no need to update the changelog.